refactor: rename sections
This PR renames the "build" and "project" sections to "Docs" and "About". These names will be more familiar to developers (and other users) and should help them get to the information they need quicker.

I know that "Concepts" is kinda "Docs" already (well the whole site is docs really) but I think the content is important enough to be a top level section.