Add "high-profile" to Filecoin section
I guess Tierion was [using it before](https://medium.com/@WayneVaughan/9f3c32f0cd9b), but I mentioned to them that we've had a lot of comms in the works for a while, so not to be offended if it looked like we were insinuating they were "low-profile". Adding "high-profile" to Filecoin should be a reasonable fix.